% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_dat.R
\name{drop_dat}
\alias{drop_dat}
\title{Remove exposures or outcomes in dat for a specific number of snps.}
\usage{
drop_dat(dat, num_snp = 5, drop_mrkeep = TRUE)
}
\arguments{
\item{dat}{dat from harmonise_data().see more \code{\link[TwoSampleMR]{harmonise_data}}}

\item{num_snp}{minimum of the number of snp}

\item{drop_mrkeep}{whether to remove snp with mr_keep parameter FALSE, the default is TRUE}
}
\value{
a data frame of the result
}
\description{
This function is designed to remove exposures or outcomes in the dat that are less than num_snp,
and the result returns the new dat after removal
}
\examples{
dat<-drop_dat(dat,num_snp=3,drop_mrkeep=TRUE)

}
