% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sbeta_res.R
\name{get_sbeta_res}
\alias{get_sbeta_res}
\title{Get res from mr() in the same beta direction}
\usage{
get_sbeta_res(
  dat,
  met = c("MR Egger", "Weighted median", "Inverse variance weighted")
)
}
\arguments{
\item{dat}{result from mr().see more \code{\link[TwoSampleMR]{mr}}}

\item{met}{participation in MR methods for comparing beta directions.Default is c("MR Egger", "Weighted median", "Inverse variance weighted")}
}
\value{
a data frame which is MR results with beta removed in different directions.
}
\description{
Removes the results of a specific MR method with a different beta direction, returning the res data frame after removal.
}
\examples{
res<-get_sbeta_res(res)

}
